requirements.txt
# ──────────────────────────────────────────────────────
# Agentic Risk Architect — Python Dependencies
# ──────────────────────────────────────────────────────
# Core Data & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0

# Market Data & News
yfinance>=0.2.30
requests>=2.31.0

# Explainability & Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Agentic / LLM
langchain>=0.1.0
anthropic>=0.20.0
crewai>=0.22.0
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0
langchain-groq>=0.1.0
litellm>=1.0.0

# Application
streamlit>=1.30.0
pydantic>=2.5.0
python-dotenv>=1.0.0
